refactor: 分离不同模块的代码
