+++ packages/s2-core/src/utils/export/copy.ts
@@ -1,15 +1,4 @@
-import {
-  escape,
-  every,
-  filter,
-  find,
-  isEmpty,
-  map,
-  max,
-  reduce,
-  repeat,
-  zip,
-} from 'lodash';
+import { escape, every, isEmpty, map, zip } from 'lodash';
 import {
   type CellMeta,
   CellTypes,
@@ -18,50 +7,59 @@ import {
   EMPTY_PLACEHOLDER,
   EXTRA_FIELD,
   InteractionStateName,
-  NODE_ID_SEPARATOR,
   type S2CellType,
-  SERIES_NUMBER_FIELD,
   VALUE_FIELD,
 } from '../../common';
-import type { Node } from '../../facet/layout/node';
 import type { SpreadSheet } from '../../sheet-type';
 import { copyToClipboard } from '../../utils/export';
 import type { ColCell, RowCell } from '../../cell';
 import {
   convertString,
-  getAllLevels,
   getColNodeFieldFromNode,
   getHeaderList,
+  getSelectedCols,
+  getSelectedRows,
   newLine,
   newTab,
 } from './method';
 import {
-  type Copyable,
   type CopyableHTML,
   type CopyableList,
   type CopyablePlain,
   CopyMIMEType,
 } from './interface';
+import { getBrushHeaderCopyable } from './copy/pivot-header-copy';
+import { processPivotSelected } from './copy/pivot-cell-copy';
+import {
+  processTableColSelected,
+  processTableRowSelected,
+} from './copy/table-copy';
 
 const getFiledFromMeta = (colIndex: number, spreadsheet: SpreadSheet) => {
   const colNode = spreadsheet
     .getColumnNodes()
     .find((col) => col.colIndex === colIndex);
+
   return getColNodeFieldFromNode(spreadsheet.isPivotMode, colNode);
 };
 
 const getHeaderNodeFromMeta = (meta: CellMeta, spreadsheet: SpreadSheet) => {
   const { rowIndex, colIndex } = meta;
+
   return [
     spreadsheet.getRowNodes().find((row) => row.rowIndex === rowIndex),
     spreadsheet.getColumnNodes().find((col) => col.colIndex === colIndex),
   ];
 };
 
-function getFormatter(spreadsheet: SpreadSheet, field: string | undefined) {
+export function getFormatter(
+  spreadsheet: SpreadSheet,
+  field: string | undefined,
+) {
   if (spreadsheet.options.interaction?.copyWithFormat) {
     return spreadsheet.dataSet.getFieldFormatter(field!);
   }
+
   return (value: DataItem) => value;
 }
 
@@ -74,6 +72,7 @@ const compatibleHideMeasureColumn = (spreadsheet: SpreadSheet) => {
   const isHideValue =
     spreadsheet.options?.style?.colCell?.hideValue &&
     spreadsheet.isValueInCols();
+
   // 被 hideMeasureColumn 隐藏的 度量(measure) 值，手动添加上。
   return isHideValue
     ? {
@@ -104,9 +103,12 @@ const getValueFromMeta = (
         colNode?.isTotals ||
         colNode?.isTotalMeasure,
     });
+
     return cell?.[VALUE_FIELD] ?? '';
   }
+
   const fieldKey = getFiledFromMeta(meta.colIndex, spreadsheet);
+
   return displayData[meta.rowIndex]?.[fieldKey!];
 };
 
@@ -117,11 +119,12 @@ const format = (
 ) => {
   const field = getFiledFromMeta(meta.colIndex!, spreadsheet);
   const formatter = getFormatter(spreadsheet, field);
+
   return formatter(getValueFromMeta(meta, displayData, spreadsheet)!);
 };
 
 // 把 DataItem[][] 矩阵转换成 CopyableItem
-const matrixPlainTextTransformer = (
+export const matrixPlainTextTransformer = (
   dataMatrix: DataItem[][],
 ): CopyablePlain => {
   return {
@@ -131,7 +134,9 @@ const matrixPlainTextTransformer = (
 };
 
 // 把 string[][] 矩阵转换成 CopyableItem
-const matrixHtmlTransformer = (dataMatrix: DataItem[][]): CopyableHTML => {
+export const matrixHtmlTransformer = (
+  dataMatrix: DataItem[][],
+): CopyableHTML => {
   function createTableData(data: DataItem[], tagName: string) {
     return data
       .map((cell) => `<${tagName}>${escape(cell as string)}</${tagName}>`)
@@ -145,6 +150,7 @@ const matrixHtmlTransformer = (dataMatrix: DataItem[][]): CopyableHTML => {
   }
 
   const body = createBody(dataMatrix, 'tr');
+
   return {
     type: CopyMIMEType.HTML,
     content: `<meta charset="utf-8"><table><tbody>${body}</tbody></table>`,
@@ -152,7 +158,7 @@ const matrixHtmlTransformer = (dataMatrix: DataItem[][]): CopyableHTML => {
 };
 
 // 生成矩阵：https://gw.alipayobjects.com/zos/antfincdn/bxBVt0nXx/a182c1d4-81bf-469f-b868-8b2e29acfc5f.png
-const assembleMatrix = (
+export const assembleMatrix = (
   rowMatrix: string[][],
   colMatrix: string[][],
   dataMatrix: string[][],
@@ -170,17 +176,20 @@ const assembleMatrix = (
     () => new Array(matrixWidth),
   );
 
-  matrix = map(matrix, (heightArr, y) => {
-    return map(heightArr, (_, x) => {
+  matrix = map(matrix, (heightArr, y) =>
+    map(heightArr, (_, x) => {
       if (x >= 0 && x < rowWidth && y >= 0 && y < colHeight) {
         return cornerMatrix?.[y]?.[x] ?? '';
       }
+
       if (x >= rowWidth && x <= matrixWidth && y >= 0 && y < colHeight) {
         return colMatrix[y][x - rowWidth];
       }
+
       if (x >= 0 && x < rowWidth && y >= colHeight && y < matrixHeight) {
         return rowMatrix[y - colHeight][x];
       }
+
       if (
         x >= rowWidth &&
         x <= matrixWidth &&
@@ -189,9 +198,10 @@ const assembleMatrix = (
       ) {
         return dataMatrix[y - colHeight][x - rowWidth];
       }
+
       return undefined;
-    });
-  });
+    }),
+  );
 
   return [matrixPlainTextTransformer(matrix), matrixHtmlTransformer(matrix)];
 };
@@ -210,9 +220,11 @@ const getSelectedCellsMeta = (cells: CellMeta[]) => {
     { row: Infinity, col: Infinity },
     { row: 0, col: 0 },
   ];
+
   // get left-top cell and right-bottom cell position
   cells.forEach((e) => {
     const { rowIndex, colIndex } = e;
+
     minCell.col = Math.min(colIndex, minCell.col);
     minCell.row = Math.min(rowIndex, minCell.row);
     maxCell.col = Math.max(colIndex, maxCell.col);
@@ -229,272 +241,11 @@ const getSelectedCellsMeta = (cells: CellMeta[]) => {
   cells.forEach((e) => {
     const { rowIndex, colIndex } = e;
     const [diffRow, diffCol] = [rowIndex - minCell.row, colIndex - minCell.col];
+
     twoDimDataArray[diffRow][diffCol] = e;
   });
+
   return twoDimDataArray;
-};
-
-const processTableColSelected = (
-  spreadsheet: SpreadSheet,
-  selectedCols: CellMeta[],
-): CopyableList => {
-  const displayData = spreadsheet.dataSet.getDisplayDataSet();
-  const columnNodes = spreadsheet.getColumnNodes();
-
-  const selectedColNodes: Node[] = selectedCols.length
-    ? (filter(columnNodes, (node) => {
-        return selectedCols.find((col) => col.id === node.id);
-      }) as Node[])
-    : columnNodes;
-
-  const dataMatrix = displayData.map((row) => {
-    return selectedColNodes.map((node) => {
-      const field = node.field;
-      const formatter = getFormatter(spreadsheet, field);
-      const value = row[field];
-      return formatter(value);
-    });
-  });
-
-  return [
-    matrixPlainTextTransformer(dataMatrix),
-    matrixHtmlTransformer(dataMatrix),
-  ];
-};
-
-const getDataMatrixByHeaderNode = (
-  leafRowNodes: Node[],
-  leafColNodes: Node[],
-  spreadsheet: SpreadSheet,
-) => {
-  return map(leafRowNodes, (rowNode) => {
-    return leafColNodes.map((colNode) => {
-      const cellData = spreadsheet.dataSet.getCellData({
-        query: {
-          ...rowNode.query,
-          ...colNode.query,
-        },
-        rowNode,
-        isTotals:
-          rowNode.isTotals ||
-          rowNode.isTotalMeasure ||
-          colNode.isTotals ||
-          colNode.isTotalMeasure,
-      });
-
-      return getFormatter(
-        spreadsheet,
-        colNode.field,
-      )(cellData?.[VALUE_FIELD] ?? '');
-    });
-  });
-};
-
-export const getPivotAllCopyData = (
-  spreadsheet: SpreadSheet,
-  leafRowNodes: Node[],
-  leafColNodes: Node[],
-): Copyable => {
-  const getCornerMatrix = (): string[][] => {
-    const { fields, meta } = spreadsheet.dataCfg;
-    const { columns = [], rows = [] } = fields;
-    const realRows = rows;
-    // const { showSeriesNumber, seriesNumberText } = spreadsheet.options;
-    // 需要考虑 serisesNumber
-    // if (showSeriesNumber) {
-    //   realRows.unshift(getDefaultSeriesNumberText(seriesNumberText));
-    // }
-    columns.push(''); // 为了对齐数值
-    // cornerMatrix 形成的矩阵为  rows.length(宽) * columns.length(高)
-    // @ts-ignore
-    const cornerMatrix: string[][] = map(columns, (col, colIndex) => {
-      return map(realRows, (row, rowIndex) => {
-        // 角头的最后一行，为行头
-        if (colIndex === columns.length - 1) {
-          return find(meta, ['field', row])?.name ?? row;
-        }
-        // 角头的最后一列，为列头
-        if (rowIndex === rows.length - 1) {
-          return find(meta, ['field', col])?.name ?? col;
-        }
-        return '';
-      });
-    });
-    // console.log(cornerMatrix, 'cornerMatrix');
-    return cornerMatrix;
-  };
-
-  const cornerMatrix = getCornerMatrix();
-  // todo-zc 的序号处理还有点麻烦，先不处理
-  const rowMatrix = map(leafRowNodes, (n) => getHeaderList(n.id));
-  const colMatrix = zip(
-    ...map(leafColNodes, (n) => getHeaderList(n.id)),
-  ) as string[][];
-  const dataMatrix = getDataMatrixByHeaderNode(
-    leafRowNodes,
-    leafColNodes,
-    spreadsheet,
-  ) as string[][];
-  return assembleMatrix(rowMatrix, colMatrix, dataMatrix, cornerMatrix);
-};
-
-function getPivotCopyData(
-  spreadsheet: SpreadSheet,
-  leafRowNodes: Node[],
-  leafColNodes: Node[],
-): CopyableList {
-  const { copyWithHeader } = spreadsheet.options.interaction!;
-  const dataMatrix = getDataMatrixByHeaderNode(
-    leafRowNodes,
-    leafColNodes,
-    spreadsheet,
-  ) as string[][];
-  // 不带表头复制
-  if (!copyWithHeader) {
-    return [
-      matrixPlainTextTransformer(dataMatrix),
-      matrixHtmlTransformer(dataMatrix),
-    ];
-  }
-  // 带表头复制
-  const rowMatrix = map(leafRowNodes, (n) => getHeaderList(n.id));
-  const colMatrix = zip(
-    ...map(leafColNodes, (n) => getHeaderList(n.id)),
-  ) as string[][];
-  return assembleMatrix(rowMatrix, colMatrix, dataMatrix);
-}
-
-const getSelectedCols = (cells: CellMeta[]) =>
-  cells.filter(({ type }) => type === CellTypes.COL_CELL);
-
-const getSelectedRows = (cells: CellMeta[]) =>
-  cells.filter(({ type }) => type === CellTypes.ROW_CELL);
-
-function processPivotSelected(
-  spreadsheet: SpreadSheet,
-  selectedCells: CellMeta[],
-): CopyableList {
-  const allRowLeafNodes = spreadsheet
-    .getRowNodes()
-    .filter((node) => node.isLeaf || spreadsheet.isHierarchyTreeType());
-  const allColLeafNodes = spreadsheet
-    .getColumnNodes()
-    .filter((node) => node.isLeaf);
-
-  let colNodes: Node[] = allColLeafNodes;
-  let rowNodes: Node[] = allRowLeafNodes;
-  const selectedColNodes = getSelectedCols(selectedCells);
-  const selectedRowNodes = getSelectedRows(selectedCells);
-
-  // selectedColNodes 选中了指定的列头，则只展示对应列头对应的数据
-  if (!isEmpty(selectedColNodes)) {
-    colNodes = selectedColNodes.reduce<Node[]>((nodes, cellMeta) => {
-      nodes.push(
-        ...allColLeafNodes.filter((node) => node.id.startsWith(cellMeta.id)),
-      );
-      return nodes;
-    }, []);
-  }
-  // selectedRowNodes 选中了指定的行头，则只展示对应行头对应的数据
-  if (!isEmpty(selectedRowNodes)) {
-    rowNodes = selectedRowNodes.reduce<Node[]>((nodes, cellMeta) => {
-      nodes.push(
-        ...allRowLeafNodes.filter((node) => node.id.startsWith(cellMeta.id)),
-      );
-      return nodes;
-    }, []);
-  }
-
-  return getPivotCopyData(spreadsheet, rowNodes, colNodes);
-}
-
-// const processPivotAllSelected = (
-//   spreadsheet: SpreadSheet,
-//   selectedCols: CellMeta[],
-// ): Copyable => {
-//   const allRowLeafNodes = spreadsheet
-//     .getRowNodes()
-//     .filter((node) => node.isLeaf);
-//   const allColLeafNodes = spreadsheet
-//     .getColumnNodes()
-//     .filter((node) => node.isLeaf);
-//
-//   const colNodes = selectedCols.length
-//     ? selectedCols.reduce<Node[]>((nodes, cellMeta) => {
-//         nodes.push(
-//           ...allColLeafNodes.filter((node) => node.id.startsWith(cellMeta.id)),
-//         );
-//         return nodes;
-//       }, [])
-//     : allColLeafNodes;
-//
-//   return getPivotAllCopyData(spreadsheet, allRowLeafNodes, colNodes);
-// };
-
-const processColSelected = (
-  spreadsheet: SpreadSheet,
-  selectedCols: CellMeta[],
-): CopyableList => {
-  if (spreadsheet.isPivotMode()) {
-    return processPivotSelected(spreadsheet, selectedCols);
-  }
-  return processTableColSelected(spreadsheet, selectedCols);
-};
-
-// const processAllSelected = (
-//   spreadsheet: SpreadSheet,
-//   selectedCols: CellMeta[],
-// ): Copyable => {
-//   if (spreadsheet.isPivotMode()) {
-//     return processPivotAllSelected(spreadsheet, selectedCols);
-//   }
-//   return processTableColSelected(spreadsheet, selectedCols);
-// };
-
-/**
- * 明细表点击行头进行复制逻辑
- * @param {SpreadSheet} spreadsheet
- * @param {CellMeta[]} selectedRows
- * @return {CopyableList}
- */
-const processTableRowSelected = (
-  spreadsheet: SpreadSheet,
-  selectedRows: CellMeta[],
-): CopyableList => {
-  const displayData = spreadsheet.dataSet.getDisplayDataSet();
-  const columnNodes = spreadsheet.getColumnNodes();
-  const matrix = map(selectedRows, (row) => {
-    const rowData = displayData[row.rowIndex];
-    // 对行内的每条数据进行格式化
-    return reduce(
-      columnNodes,
-      (acc: string[], colNode: Node) => {
-        const key = colNode?.field;
-        const value = rowData[key];
-        const noFormatting = !colNode || key === SERIES_NUMBER_FIELD;
-        if (noFormatting) {
-          return acc;
-        }
-        const formatterVal = convertString(
-          getFormatter(spreadsheet, colNode?.field)(value),
-        ) as string;
-        acc.push(formatterVal);
-        return acc;
-      },
-      [],
-    );
-  });
-  return [matrixPlainTextTransformer(matrix), matrixHtmlTransformer(matrix)];
-};
-
-const processRowSelected = (
-  spreadsheet: SpreadSheet,
-  selectedRows: CellMeta[],
-): CopyableList => {
-  if (spreadsheet.isPivotMode()) {
-    return processPivotSelected(spreadsheet, selectedRows);
-  }
-  return processTableRowSelected(spreadsheet, selectedRows);
 };
 
 /**
@@ -511,11 +262,9 @@ const getDataMatrixByDataCell = (
 ): CopyableList => {
   const { copyWithHeader } = spreadsheet.options.interaction!;
 
-  const dataMatrix = map(cellMetaMatrix, (cellsMeta) => {
-    return map(cellsMeta, (it) =>
-      convertString(format(it, displayData, spreadsheet)),
-    );
-  }) as string[][];
+  const dataMatrix = map(cellMetaMatrix, (cellsMeta) =>
+    map(cellsMeta, (it) => convertString(format(it, displayData, spreadsheet))),
+  ) as string[][];
 
   if (!copyWithHeader) {
     return [
@@ -528,6 +277,7 @@ const getDataMatrixByDataCell = (
   const colMatrix = zip(
     ...map(cellMetaMatrix[0], (cellMeta) => {
       const colId = cellMeta.id.split(EMPTY_PLACEHOLDER)?.[1] ?? '';
+
       return getHeaderList(colId);
     }),
   ) as string[][];
@@ -535,133 +285,45 @@ const getDataMatrixByDataCell = (
   // 通过第一列来获取行头信息
   const rowMatrix = map(cellMetaMatrix, (cellsMeta) => {
     const rowId = cellsMeta[0].id.split(EMPTY_PLACEHOLDER)?.[0] ?? '';
+
     return getHeaderList(rowId);
   });
 
   return assembleMatrix(rowMatrix, colMatrix, dataMatrix);
 };
 
-/**
- * 过滤出 intersection cell 中所有叶子节点的 cellMeta
- * @param interactedCells
- * @param maxLevel
- * @returns {CellMeta[]}
- */
-function getLastLevelCells(
-  interactedCells: RowCell[] | ColCell[],
-  maxLevel: number,
-) {
-  return filter(interactedCells, (cell: RowCell | ColCell) => {
-    const meta = cell.getMeta();
-    const isLastLevel = meta.level === maxLevel;
-    const isLastTotal = meta.isTotals && isEmpty(meta.children);
-    return isLastLevel || isLastTotal;
-  });
-}
-
-/**
- * 获取表头圈选后的 header cells 值矩阵
- * @param lastLevelCells
- * @param maxLevel
- * @param allLevel
- * @returns {string[][]}
- */
-function getHeaderCellMatrix(
-  lastLevelCells: Array<RowCell | ColCell>,
-  maxLevel: number,
-  allLevel: Set<number>,
-): string[][] {
-  return map(lastLevelCells, (cell: RowCell | ColCell) => {
-    const meta = cell.getMeta();
-    const { id, value, isTotals, level } = meta;
-    let cellId = id;
-    // 为总计小计补齐高度
-    if (isTotals && level !== maxLevel) {
-      cellId = id + NODE_ID_SEPARATOR + repeat(value, maxLevel - level);
-    }
-    return getHeaderList(cellId, allLevel.size);
-  });
-}
-
-function getBrushHeaderCopyable(
-  interactedCells: RowCell[] | ColCell[],
-): CopyableList {
-  // 获取圈选的层级有哪些
-  const allLevels = getAllLevels(interactedCells);
-  const maxLevel = max(Array.from(allLevels)) ?? 0;
-  // 获取最后一层的 cell
-  const lastLevelCells = getLastLevelCells(interactedCells, maxLevel) as Array<
-    RowCell | ColCell
-  >;
-
-  // 拼接选中行列头的内容矩阵
-  const isCol = interactedCells[0].cellType === CellTypes.COL_CELL;
-  let cellMatrix = getHeaderCellMatrix(lastLevelCells, maxLevel, allLevels);
-
-  // 如果是列头，需要转置
-  if (isCol) {
-    cellMatrix = zip(...cellMatrix) as string[][];
+const processColSelected = (
+  spreadsheet: SpreadSheet,
+  selectedCols: CellMeta[],
+): CopyableList => {
+  if (spreadsheet.isPivotMode()) {
+    return processPivotSelected(spreadsheet, selectedCols);
   }
-  return [
-    matrixPlainTextTransformer(cellMatrix),
-    matrixHtmlTransformer(cellMatrix),
-  ];
+
+  return processTableColSelected(spreadsheet, selectedCols);
+};
+
+const processRowSelected = (
+  spreadsheet: SpreadSheet,
+  selectedRows: CellMeta[],
+): CopyableList => {
+  if (spreadsheet.isPivotMode()) {
+    return processPivotSelected(spreadsheet, selectedRows);
+  }
+
+  return processTableRowSelected(spreadsheet, selectedRows);
+};
+
+function getIsBrushHeader(interactedCells: S2CellType[]) {
+  return isEmpty(interactedCells)
+    ? false
+    : every(
+        interactedCells,
+        (cell) =>
+          cell.cellType === CellTypes.ROW_CELL ||
+          cell.cellType === CellTypes.COL_CELL,
+      );
 }
-
-// const tilePivotData = (
-//   data: any,
-//   columnOrdered: CustomHeaderField[],
-//   defaultDataValue?: string,
-// ): Array<string> => {
-//   // @ts-ignore
-//   return map(columnOrdered, (field) => data[field] ?? defaultDataValue);
-// };
-
-// const getDataByRowData = (
-//   spreadsheet: SpreadSheet,
-//   rowData: RowData,
-// ): Copyable => {
-//   const {
-//     options: { placeholder },
-//     dataCfg: {
-//       fields: { rows = [], columns = [], values = [] },
-//     },
-//   } = spreadsheet;
-//   const defaultDataValue = getEmptyPlaceholder(spreadsheet, placeholder);
-//   const column = spreadsheet.getColumnLeafNodes();
-//   let datas: string[][] = [];
-//
-//   if (spreadsheet.isTableMode()) {
-//     const columnWithoutSeriesNumber = filter(
-//       column,
-//       (node) => node.field !== SERIES_NUMBER_FIELD,
-//     );
-//     // 按列头顺序复制
-//     datas = map(rowData, (rowDataItem) => {
-//       return map(
-//         columnWithoutSeriesNumber,
-//         // @ts-ignore
-//         (node) => rowDataItem?.[node.field] ?? defaultDataValue,
-//       );
-//     });
-//   } else if (spreadsheet.isPivotMode()) {
-//     // 透视表的数据加上行头、列头才有意义，这里会以行头、列头、数据值的顺序将每一个单元格构造成一行
-//     const columnOrdered = [...rows, ...columns, ...values];
-//     const rowDataFlatten = flattenDeep(rowData as unknown as Array<RowData>);
-//     // 去掉小计
-//     const rowDataFlattenWithoutTotal = rowDataFlatten.filter((data) =>
-//       [...rows, ...columns].every((field) => !isNil(data[field as string])),
-//     );
-//     datas = reduce(
-//       rowDataFlattenWithoutTotal,
-//       (ret: unknown[], data) => {
-//         return [...ret, tilePivotData(data, columnOrdered, defaultDataValue)];
-//       },
-//       [],
-//     ) as string[][];
-//   }
-//   return matrixPlainTextTransformer(datas);
-// };
 
 function getDataCellCopyable(
   spreadsheet: SpreadSheet,
@@ -696,34 +358,28 @@ function getDataCellCopyable(
         },
       ];
     }
+
     // normal selected
     const selectedCellsMeta = getSelectedCellsMeta(cells);
-    // console.log('selectedCellsMeta', selectedCellsMeta);
-    // const { currentRow } = spreadsheet.interaction.getSelectedCellHighlight();
-    // if (currentRow) {
-    //   const rowData = orderBy(cells, 'rowIndex', 'asc').map((cell) =>
-    //     spreadsheet.dataSet.getRowData(cell),
-    //   );
-    //   data = getDataByRowData(spreadsheet, rowData as unknown as RowData);
-    // } else
+
+    /*
+     * console.log('selectedCellsMeta', selectedCellsMeta);
+     * const { currentRow } = spreadsheet.interaction.getSelectedCellHighlight();
+     * if (currentRow) {
+     *   const rowData = orderBy(cells, 'rowIndex', 'asc').map((cell) =>
+     *     spreadsheet.dataSet.getRowData(cell),
+     *   );
+     *   data = getDataByRowData(spreadsheet, rowData as unknown as RowData);
+     * } else
+     */
     data = getDataMatrixByDataCell(
       selectedCellsMeta,
       displayData as Data[],
       spreadsheet,
     );
   }
-  return data;
-}
 
-function getIsBrushHeader(interactedCells: S2CellType[]) {
-  return isEmpty(interactedCells)
-    ? false
-    : every(interactedCells, (cell) => {
-        return (
-          cell.cellType === CellTypes.ROW_CELL ||
-          cell.cellType === CellTypes.COL_CELL
-        );
-      });
+  return data;
 }
 
 // 刷选复制使用
@@ -745,5 +401,6 @@ export const getSelectedData = (spreadsheet: SpreadSheet): CopyableList => {
   if (data) {
     copyToClipboard(data);
   }
+
   return data!;
 };
